requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==7.1.0
phonenumberslite==8.13.16